writewise/frontend/requirements.txt
# Core dependencies
streamlit==1.24.0
requests==2.31.0
python-dotenv==1.0.0


